Simon domasno 3 facilities info

+++ Domasna3/src/main/java/com/example/diansspring/model/MapCoordinates.java
@@ -6,11 +6,11 @@ import java.util.Objects;
 
 @Data
 public class MapCoordinates {
-    private double lat;
-    private double lng;
-    private double zoom;
+    private float lat;
+    private float lng;
+    private float zoom;
 
-    public MapCoordinates(Double lat, Double lng, Double zoom) {
+    public MapCoordinates(Float lat, Float lng, Float zoom) {
         this.lat = Objects.equals(lat, null) ? -1 : lat;
         this.lng = Objects.equals(lng, null) ? -1 : lng;
         this.zoom = Objects.equals(zoom, null) ? -1 : zoom;
